Adds functioning curl request for PATCH
Successfully able to make a curl request to PATCH surveys. Currently
updating the "answers" field. Realized this is going to be a problem
in that it overwrites the answer rather than appending to end of the
"answers" array. Next steps are to change this PATCH to a POST request
to surveys resource to create a new answer and append that to end of
"answers" array. We will make PATCH request for changing title of
survey.