Begins working on "create answer" curl
Update routes, controller, scripts to create a new answer to a survey.
We do not have a functioning curl request for this. We are still unclear
as to whether we can POST an answer to the surveys resource and append
the new answer to the "answers" array OR if we have to do a PATCH to
surveys resource and append new answer to "answers" array. Currently,
a PATCH overwrites the previous answer which is not what we want. Other
possible solution would be to create separate answers collection and
reference survey id to link them.